create ambient_utils to clean ambient
